partially refactored User

+++ src/main/java/it/hospital/project/appusers/UserController.java
@@ -3,7 +3,14 @@ package it.hospital.project.appusers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 
 @RestController
@@ -25,15 +32,14 @@ public class UserController {
 
 
     @PostMapping
-    ResponseEntity<?> addUser(@RequestBody  User user) {
+    ResponseEntity<?> addUser(@RequestBody User user) {
         return new ResponseEntity<>(userService.addUser(user), HttpStatus.OK);
     }
 
     @DeleteMapping("/{userId}")
-    ResponseEntity<?> deleteUser(@PathVariable int userId) {
+    ResponseEntity<?> deleteUser(@PathVariable Long userId) {
         userService.deleteUser(userId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
-
 
 }
